Refactor, serialize jsonld on build.
