--- contexts/example-v1.json
@@ -1,6 +1,0 @@
-{
-  "@context": {
-    "id": "@id",
-    "type": "@type"
-  }
-}
