+++ contexts/example-v1.jsonld
@@ -1,0 +1,7 @@
+{
+  "@context": {
+    "id": "@id",
+    "type": "@type",
+    "@protected": true
+  }
+}
