+++ js/constants.js
@@ -4,6 +4,7 @@
 'use strict';
 
 module.exports = {
+  CONTEXT_FILENAME: 'example-v1.jsonld',
   CONTEXT_URL: 'https://w3id.org/example/v1',
   // value between 0x0 and 0x7FFF for globally registered term codec values
   // or >= 0x8000 for app-specific local terms
