+++ js/context.js
@@ -1,0 +1,13 @@
+/*!
+ * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
+ */
+'use strict';
+
+module.exports = {
+  '@context': {
+    id: '@id',
+    type: '@type',
+    '@protected': true
+
+  }
+};
