--- js/documentLoader.js
@@ -1,21 +1,0 @@
-/*!
- * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
- */
-'use strict';
-
-const {CONTEXT_URL} = require('./constants');
-const context = require('../contexts/example-v1.json');
-
-module.exports = {
-  documentLoader(url) {
-    if(url !== CONTEXT_URL) {
-      throw new Error(`Loading document "${url}" is not allowed.`);
-    }
-
-    return {
-      contextUrl: null,
-      document: context,
-      documentUrl: url
-    };
-  }
-};
