+++ js/index.js
@@ -3,13 +3,12 @@
  */
 'use strict';
 
-const context = require('../contexts/example-v1.json');
+const context = require('./context');
 const constants = require('./constants');
-const {documentLoader} = require('./documentLoader');
 const {CONTEXT_URL, CBORLD_VALUE} = constants;
 
 const contexts = new Map();
-contexts.set(constants.CONTEXT_URL, context);
+contexts.set(CONTEXT_URL, context);
 
 const appContextMap = new Map();
 appContextMap.set(CONTEXT_URL, CBORLD_VALUE);
@@ -18,5 +17,6 @@ module.exports = {
   constants,
   contexts,
   appContextMap,
-  documentLoader
+  CONTEXT_URL,
+  CONTEXT: context
 };
