+++ package.json
@@ -2,12 +2,10 @@
   "name": "context-template",
   "version": "0.0.1",
   "description": "Context template repository.",
-  "main": "js",
-  "module": "./dist/context.esm.js",
+  "main": "js/index.js",
   "files": [
     "contexts",
     "dist/context.js",
-    "dist/context.esm.js",
     "js"
   ],
   "repository": {
@@ -51,12 +49,13 @@
     "webpack-cli": "^3.3.12"
   },
   "scripts": {
-    "build": "npm run clean && webpack && rollup -c",
-    "clean": "rm -rf dist/*",
+    "build": "npm run clean && npm run serialize && webpack && rollup -c",
+    "clean": "rm -rf dist/* && rm -rf contexts/*",
     "lint": "eslint js test",
     "prepare": "npm run build",
+    "serialize": "node js/serialize.js",
     "test": "npm run lint && npm run build && npm run test-node && npm run test-karma",
     "test-karma": "karma start test/karma.conf.js",
-    "test-node": "cross-env NODE_ENV=test mocha --preserve-symlinks test/**/*.spec.js"
+    "test-node": "cross-env NODE_ENV=test mocha --preserve-symlinks test/*.spec.js"
   }
 }
